refactor(presentation): Switched screens for views to handle PageView swapping better
