+++ lib/config/router/router.dart
@@ -1,21 +1,19 @@
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:go_router/go_router.dart';
+import 'package:libra_ui/domain/models/auth/auth_data.dart';
 import 'package:libra_ui/presentation/providers/auth/auth_provider.dart';
-import 'package:libra_ui/presentation/screens/account/settings/settings_screen.dart';
-import 'package:libra_ui/presentation/screens/account/transaction/transaction_history_screen.dart';
-import 'package:libra_ui/presentation/screens/account/transfer/transfer_screen.dart';
-import 'package:libra_ui/presentation/screens/screens.dart';
 import 'package:libra_ui/presentation/screens/auth/login_screen.dart';
 import 'package:libra_ui/presentation/screens/auth/register_screen.dart';
+import 'package:libra_ui/presentation/screens/home/home_screen.dart';
 
 // It's good practice to define route names as constants
 class AppRoutes {
-  static const String home = '/';
+  static const String home = '/home';
   static const String login = '/login';
   static const String register = '/register';
   static const String transactions = '/transactions';
-  static const String settings = '/settings';
   static const String transfer = '/transfer';
+  static const String settings = '/settings';
 
   static const List<String> values = [
     home,
@@ -31,15 +29,19 @@ final routerProvider = Provider<GoRouter>((ref) {
   final authState = ref.watch(authRepositoryProvider);
 
   final isAuthenticated = authState.token.isNotEmpty;
-
-  return GoRouter(
-    initialLocation: AppRoutes.home,
+  final homeRoute = GoRoute(
+    path: '${AppRoutes.home}/:page',
+    name: AppRoutes.home,
+    builder: (context, state) {
+      final pageIndex = _getRouteIndex(state);
+      if (pageIndex > 2 || pageIndex < 0) return const HomeScreen();
+      return HomeScreen(pageIndex: pageIndex);
+    },
+  );
+  final router = GoRouter(
+    initialLocation: isAuthenticated ? '${AppRoutes.home}/0' : AppRoutes.login,
     routes: [
-      GoRoute(
-        path: AppRoutes.home,
-        name: AppRoutes.home,
-        builder: (context, state) => const HomeScreen(),
-      ),
+      homeRoute,
       GoRoute(
         path: AppRoutes.login,
         name: AppRoutes.login,
@@ -49,21 +51,6 @@ final routerProvider = Provider<GoRouter>((ref) {
         path: AppRoutes.register,
         name: AppRoutes.register,
         builder: (context, state) => const RegisterScreen(),
-      ),
-      GoRoute(
-        path: AppRoutes.transactions,
-        name: AppRoutes.transactions,
-        builder: (context, state) => const TransactionsScreen(),
-      ),
-      GoRoute(
-        path: AppRoutes.transfer,
-        name: AppRoutes.transfer,
-        builder: (context, state) => const TransferScreen(),
-      ),
-      GoRoute(
-        path: AppRoutes.settings,
-        name: AppRoutes.settings,
-        builder: (context, state) => const SettingsScreen(),
       ),
     ],
     redirect: (context, state) {
@@ -91,4 +78,14 @@ final routerProvider = Provider<GoRouter>((ref) {
       return null;
     },
   );
+  // Listen to auth changes to redirect to login on logout
+  ref.listen<AuthData>(authRepositoryProvider, (previous, next) {
+    if (previous?.token.isNotEmpty == true && next.token.isEmpty) {
+      router.go(AppRoutes.login);
+    }
+  });
+  return router;
 });
+
+int _getRouteIndex(GoRouterState state) =>
+    int.tryParse(state.pathParameters['page'] ?? '0') ?? 0;
