--- lib/presentation/screens/account/transaction/transaction_history_screen.dart
@@ -1,38 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:libra_ui/config/theme/libra_colors.dart';
-import 'package:libra_ui/presentation/providers/account/account_provider.dart';
-import 'package:libra_ui/presentation/widgets/home/transaction_history.dart';
-import 'package:libra_ui/presentation/widgets/shared/navbar.dart';
-
-class TransactionsScreen extends ConsumerWidget {
-  const TransactionsScreen({super.key});
-
-  @override
-  Widget build(BuildContext context, WidgetRef ref) {
-    final transactions = ref.watch(accountProvider).transactions;
-    return Scaffold(
-      backgroundColor: LibraColors.scaffoldBackground,
-      body: SafeArea(
-        child: Column(
-          children: <Widget>[
-            Expanded(
-              child: Padding(
-                padding: const EdgeInsets.all(16.0),
-                child: TransactionHistory(
-                  cardBackgroundColor: LibraColors.cardBackground,
-                  accentColorTeal: LibraColors.accentTeal,
-                  primaryTextColor: LibraColors.primaryText,
-                  secondaryTextColor: LibraColors.secondaryText,
-                  transactions: transactions,
-                  limit: transactions.length,
-                ),
-              ),
-            ),
-          ],
-        ),
-      ),
-      bottomNavigationBar: NavBar.common(),
-    );
-  }
-}
