+++ lib/presentation/screens/auth/login_screen.dart
@@ -6,7 +6,6 @@ import 'package:libra_ui/config/theme/libra_colors.dart';
 import 'package:libra_ui/presentation/providers/auth/auth_provider.dart';
 import 'package:libra_ui/presentation/providers/auth/auth_message_provider.dart';
 import 'package:libra_ui/presentation/widgets/auth/auth_form.dart';
-// TODO: Import a shared button widget if available, or create one.
 
 class LoginScreen extends ConsumerStatefulWidget {
   const LoginScreen({super.key});
