+++ lib/presentation/screens/screens.dart
@@ -1,2 +1,4 @@
 // Bucket file for screens
 export 'home/home_screen.dart';
+export 'auth/login_screen.dart';
+export 'auth/register_screen.dart';
