+++ lib/presentation/views/account/transfer/transfer.dart
@@ -1,0 +1,4 @@
+export 'transfer_view.dart';
+export 'transfer_amount_view.dart';
+export 'transfer_confirmation_view.dart';
+export 'transfer_loading_view.dart';
