+++ lib/presentation/views/views.dart
@@ -1,0 +1,4 @@
+export 'account/transfer/transfer.dart';
+export 'account/transaction/transaction_history_view.dart';
+export 'account/settings/settings_view.dart';
+export 'home/home_view.dart';
